requirements.txt
asgiref==3.7.2
Django==4.2.4
django-jazzmin==2.6.0
django-multiselectfield==0.1.12
djangorestframework==3.14.0
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
